try double

+++ project1/code/Min.c
@@ -1,7 +1,7 @@
 #include<linux/kernel.h>
 #include<linux/linkage.h>
 
-asmlinkage long sys_Min(const long a, const long b)
+asmlinkage double sys_Min(const double a, const double b)
 {
 	if(a<b)
 		return a;
